feat

+++ app/Http/Controllers/ComparativeOutflowsDashboardController.php
@@ -46,7 +46,7 @@ class ComparativeOutflowsDashboardController extends Controller
         $adminUser = \App\Models\User::where('team_id', $team_id)
             ->role('Admin')
             ->first();
-        $dollarPrice = $adminUser?->dollar_price ?? 970;
+        $dollarPrice = (float) ($adminUser?->dollar_price ?? 970);
 
         // Obtener información de la temporada
         $season = Season::with('month')->find($season_id);
