+++ app/Http/Controllers/InvestmentDashboardController.php
@@ -26,7 +26,7 @@ class InvestmentDashboardController extends Controller
         $adminUser = \App\Models\User::where('team_id', $team_id)
             ->role('Admin')
             ->first();
-        $dollarPrice = $adminUser?->dollar_price ?? 970;
+        $dollarPrice = (float) ($adminUser?->dollar_price ?? 970);
 
         // Obtener información de la temporada
         $season = Season::with('month')->find($season_id);
